Abstracted netdevil patcher into internal package
